+++ packages/@expo/cli/src/utils/env.ts
@@ -90,8 +90,8 @@ class Env {
   }
 
   /** Enable auto server root detection for Metro. This will change the server root to the workspace root. */
-  get EXPO_USE_METRO_SERVER_ROOT(): boolean {
-    return boolish('EXPO_USE_METRO_SERVER_ROOT', false);
+  get EXPO_USE_METRO_WORKSPACE_ROOT(): boolean {
+    return boolish('EXPO_USE_METRO_WORKSPACE_ROOT', false);
   }
 
   /**